+++ src/components/AppNewsletterSection.tsx
