+++ src/components/AppNoticeSection.tsx
