+++ src/components/AuthorDetailsCard.tsx
