+++ src/components/CollectionsFooter.tsx
