+++ src/components/VotingComponentArticle.tsx
